Compatibility layer between python-sasl and pure-sasl
python-sasl is used as main SASL library, if it's not there, pure-sasl
is used.
